// Frontend/lib/types/lesson-hierarchy.types.ts
export interface ModuleOption {
	id: number;
	name: string;
}

export interface SubjectOption {
	id: number;
	name: string;
	type: string;
}

export interface ChapterOption {
	id: number;
	name: string;
}

export interface LessonOption {
	id: number;
	name: string;
}

export interface SubjectResponse {
	id: number;
	name: string;
	type: string;
}

export interface ChapterResponse {
	id: number;
	name: string;
}

export interface LessonResponse {
	id: number;
	name: string;
}